tools/downloader: split off the concurrency-related functionality into a module
Leave `JobContext` and `DirectOutputContext` where they are now, since
they aren't directly related to concurrency. Also, Reporter relies on them.

Turn QueuedOutputContext and JobWithQueuedOutput private, since they aren't
supposed to be used directly.
